requirements.txt
flask
python-dotenv
git+https://github.com/supabase-community/supabase-py.git
